remove grpc related stuff from public headers
