+++ src/Private/GrpcHelpers.cpp
@@ -6,9 +6,19 @@
  */
 
 #include "GrpcHelpers.h"
+#include "ai/inworld/studio/v1alpha/characters.pb.h"
 
 Inworld::GrpcHelper::CharacterInfo Inworld::GrpcHelper::CreateCharacterInfo(const InworldV1alpha::Character& GrpcCharacter)
 {
 	Inworld::GrpcHelper::CharacterInfo Info(GrpcCharacter);
 	return Info;
 }
+
+Inworld::GrpcHelper::CharacterInfo::CharacterInfo(const InworldV1alpha::Character& GrpcCharacter)
+	: _Name(GrpcCharacter.name())
+	, _RpmModelUri(GrpcCharacter.default_character_assets().rpm_model_uri())
+	, _RpmImageUri(GrpcCharacter.default_character_assets().rpm_image_uri())
+	, _RpmPortraitUri(GrpcCharacter.default_character_assets().rpm_image_uri_portrait())
+	, _RpmPostureUri(GrpcCharacter.default_character_assets().rpm_image_uri_posture())
+	, _bMale(GrpcCharacter.default_character_description().pronoun() == InworldV1alpha::Character_CharacterDescription_Pronoun_PRONOUN_MALE)
+{}
